requirements.txt
flask
plaid-python>=20.0.0
python-dotenv
requests